Scala-pickling: added support for recursive types via implicit picklers
This is the result of Heather and I working out the proposal of Eugene and Denys
to store the generated picklers in implicit values, so that they can be found
by implicit search in case they are needed, for example, in the case of
recursive types.

A welcome side effect is the removal of a map for looking up the right picklers
to pickle field values.