disables osgi packaging and tests
Quasiquote implementation uses quasiquotes, which generate unappliers
into packages having dollars in their names. This is a problem for an
OSGi bridge we use to run some tests: https://github.com/bndtools/bnd/issues/140.

The aforementioned issue is fixed, but I haven't been able to find
a published version of the bridge with the fix. Therefore I'm disabling
the tests for the time being.