implements the refactoring discussed on Mar 16
1) Picklers, unpicklers and pickle formats have been reimplemented using
the new notions of PickleBuilder and PickleReader.

2) Synthetic picklers and unpicklers are now introduced into the top-level
scope rather than generated locally. This means that nested dispatches
generated into their bodies might be potentially imprecise. We discussed
that yesterday and came to the conclusion that we can sacrifice this
for the sake of convenience.

3) Pickle dispatch now uses Class[_], because that's the best we can
reliably have at runtime. After we dispatch to a pickler using an erasure
of a type, the pickler can apply arbitrary strategies to figure out the
actual values of type arguments (not implemented yet).

4) Unpickle dispatch uses types, not erasures, because that's what we
serialize to pickles. Some measures have been taken to make this friendly
to polymorphic classes (e.g. see an existentialAbstraction call in Tools).

5) Integer and string picklers/unpicklers are now implemented as custom
picklers/unpicklers. This has brought some grief w.r.t implicit search,
which I have fixed in the previous commit. Implementing custom stuff
also gave some insight w.r.t the current design - we need to discuss the
findings.

6) FieldIR now has flags, which support queries to things like isPublic,
isVar, isReifiedParam, isErasedParam. !isPublic, isVar and isErasedParam
are marked as not yet implemented.