don't share arrays
Sharing arrays belongs to the same category of impossible as sharing
lists or vectors. It is very time-consuming to find whether these things
are equal, and in case of arrays this is beyond normal hashCode().

Therefore I'm just either leaving this for future work when you guys
figure out what datastructure you're planning to use or forgetting about
this for good.