Changed JSON field "tpe" to "$type" (issue #64)
This makes it consistent with the naming of the "$ref" field and makes collisions with ordinary (unquoted) Scala fields much less likely. This is consistent with the Scala Language Specification section 1.1: "The ‘$’ character is reserved for compiler-synthesized identiﬁers. User programs
should not deﬁne identiﬁers which contain ‘$’ characters."