add directSubclasses annotation
This gives people a way to control or work around the limitations
of Scala's knownDirectSubclasses method. It also allows generating
a static pickler for non-sealed traits or classes, where the pickler
of course will fail for new subtypes not included in the annotation.
This commit also improves the error message when pickling an unknown
subtype (with this commit it throws PicklingException rather than
MatchError).