Quick and dirty prototype of using ASM to read java class files.
It turns out the scala compiler doens't load private fields + transient annotations in certain modes,
so if we're going to look at java code, we need to go deeper.  Using the built-in ASM in scala compiler
will work for macros.   We'll rely on exisitng code for runtime pickling,a nd just get the macros to be more
accurate.