Deprecate sharing.
Sharing is not proven to be safe.  The only way sharing can work is if we always use Unsafe to
instantiate new instances and reflectively inject members.   This technique works on the JVM,
but not Android, and additionally we want users to be able to opt out of any runtime
reflection.